chore(#123): integrate all scenarios into one file
